third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-stretch-expected.txt

This is a testharness.js-based test.
PASS Can set 'font-stretch' to CSS-wide keywords
PASS Can set 'font-stretch' to var() references
FAIL Can set 'font-stretch' to the 'normal' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'ultra-condensed' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'extra-condensed' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'condensed' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'semi-condensed' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'semi-expanded' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'expanded' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'extra-expanded' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to the 'ultra-expanded' keyword assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
FAIL Can set 'font-stretch' to a percent assert_approx_equals: expected -3.14 +/- 0.000001 but got -3
PASS Setting 'font-stretch' to a length throws TypeError
PASS Setting 'font-stretch' to a time throws TypeError
PASS Setting 'font-stretch' to an angle throws TypeError
PASS Setting 'font-stretch' to a flexible length throws TypeError
PASS Setting 'font-stretch' to a number throws TypeError
PASS Setting 'font-stretch' to a position throws TypeError
PASS Setting 'font-stretch' to a URL throws TypeError
PASS Setting 'font-stretch' to a transform throws TypeError
Harness: the test ran to completion.
